INT: add documentation for new `Show macro expansion` intentions
Also improve test for intention documentation to check documentation existence
of all intentions provided by plugin
