+++ src/main/resources/intentionDescriptions/RsShowRecursiveMacroExpansionIntention/description.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+This intention shows popup with result of full macro expansion.
+</body>
+</html>
