+++ src/main/resources/intentionDescriptions/RsShowSingleStepMacroExpansionIntention/description.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+This intention shows popup with result of single step macro expansion (without expansion of nested macro invocations).
+</body>
+</html>
