+++ src/test/kotlin/org/rust/ide/intentions/RsIntentionTestBase.kt
@@ -13,19 +13,6 @@ import org.rust.lang.RsTestBase
 import org.rust.openapiext.Testmark
 
 abstract class RsIntentionTestBase(val intention: IntentionAction) : RsTestBase() {
-    fun `test intention has documentation`() {
-        val directory = "intentionDescriptions/${intention.javaClass.simpleName}"
-        val files = listOf("before.rs.template", "after.rs.template", "description.html")
-        for (file in files) {
-            val text = getResourceAsString("$directory/$file")
-                ?: fail("No inspection description for ${intention.javaClass}.\n" +
-                "Add ${files.joinToString()} to src/main/resources/$directory")
-
-            if (file.endsWith(".html")) {
-                checkHtmlStyle(text)
-            }
-        }
-    }
 
     protected fun doAvailableTest(@Language("Rust") before: String, @Language("Rust") after: String) {
         InlineFile(before.trimIndent()).withCaret()
